+++ src/app/app/(with_back_button)/event/create/new/page.tsx
@@ -15,11 +15,25 @@ export default async function CreateEventPage() {
     )
     .select('description', 'label', 'ec.id as id');
 
+  const thresholds = await db({ et: tablenames.event_threshold })
+    .leftJoin(
+      db
+        .select('event_threshold_id', 'description')
+        .from(tablenames.event_threshold_description)
+        .as('etd'),
+      'etd.event_threshold_id',
+      'et.id'
+    )
+    .select('et.id', 'et.label', 'etd.description');
+
   return (
     <div className='flex flex-col px-default w-full flex-1 justify-center items-center'>
       <FormContainer>
         <h2>Create event</h2>
-        <CreateEventForm categories={categories} />
+        <CreateEventForm
+          categories={categories}
+          thresholds={thresholds}
+        />
       </FormContainer>
     </div>
   );